fix(dockerfile, docker-compose.yaml): add wait for db and redis to server container
When starting the server via docker-compose currently the server starts up faster than the postgres
instance. This causes the server app to crash. The fix now adds a wait script that pings the
database and the cache containers unit they properly start up. This delays the server boot until all
deps are started.