feat(helm chart): secrets can be referenced from different kubernetes Secret resources
Currently secrets have to be referenced from a single kubernetes Secret resource (default name
'server-vars').  This PR allows each secret to be loaded from a separate kubernetes Secret.  If
values for individual secrets are not provided, it defaults to the previous single kubernetes
resource.  This single kubernetes secret should now be considered deprecated in favour of individual
references.